CLN: Move version file to `xtgeo.common.version`
This makes mypy resolution more straight forward when the version is
used in deprecation notices.
