+++ pyproject.toml
@@ -13,7 +13,7 @@ build-backend = "scikit_build_core.build"
 cmake.minimum-version = "3.15.3"
 cmake.verbose = true
 metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
-sdist.include = ["src/xtgeo/_theversion.py"]
+sdist.include = ["src/xtgeo/common/version.py"]
 wheel.install-dir = "xtgeo"
 
 [project]
@@ -111,7 +111,7 @@ exclude = '''
   | buck-out
   | build
   | dist
-  | _theversion.py
+  | version.py
   # The following are specific to Black, you probably don't want those.
   | blib2to3
   | tests/data
@@ -155,4 +155,4 @@ markers = [
 doctest_optionflags = "ELLIPSIS"
 
 [tool.setuptools_scm]
-write_to = "src/xtgeo/_theversion.py"
+write_to = "src/xtgeo/common/version.py"
