+++ src/xtgeo/__init__.py
@@ -10,9 +10,7 @@ import timeit
 import warnings
 
 try:
-    from ._theversion import version
-
-    __version__ = version
+    from xtgeo.common.version import __version__, version
 except ImportError:
     __version__ = "0.0.0"
